fixed a small formatting issue in celPcCommon::AddProperty boundary check report that was provoking a lot of warnings.

+++ include/celtool/stdpcimp.h
@@ -139,11 +139,11 @@ protected:
       csRef<iReporter> rep = csQueryRegistry<iReporter>(object_reg);
       if (rep)
         rep->ReportError("crystalspace.cel.physicallayer",
-            "celPcCommon::AddProperty out of bounds %d >= %d!",
-            idx,propholder->propertycount);
+            "celPcCommon::AddProperty out of bounds %u >= %u!",
+            (unsigned)idx,(unsigned)propholder->propertycount);
       else
-        printf("Error: celPcCommon::AddProperty out of bounds %d >= %d!",
-            idx,propholder->propertycount);
+        printf("Error: celPcCommon::AddProperty out of bounds %u >= %u!",
+            (unsigned)idx,(unsigned)propholder->propertycount);
       return;
     }
     if (propdata == 0)
